requirements.txt
streamlit
requests
gitpython
Pillow
google-generativeai
# Add any other dependencies used by your agents below
